requirements.txt
asgiref==3.3.1
autopep8==1.5.5
dj-database-url==0.5.0
Django==3.1.6
django-heroku==0.3.1
gunicorn==20.0.4
numpy==1.19.5
pandas==1.1.5
Pillow==8.1.0
psycopg2==2.7.5
psycopg2-binary==2.8.6
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
